requirements.txt
--find-links https://download.pytorch.org/whl/torch_stable.html
numpy>=1.14.0
torch==1.9.0+cu111
gym==0.19.0
tensorboardX==1.2
pyyaml==5.4.1
matplotlib
seaborn==0.11.1
gym_minigrid
mujoco-py<2.1,>=2.0
